[REF] dbfilter_from_header: Use post_load hook to apply patch
It avoids apply patch if the module is not installed
