+++ dbfilter_from_header/__init__.py
@@ -1,2 +1,2 @@
 
-from . import override
+from .hooks import post_load
