+++ dbfilter_from_header/__manifest__.py
@@ -25,4 +25,5 @@
     "external_dependencies": {
         'python': [],
     },
+    "post_load": "post_load",
 }
